realm db,view list,update

+++ app/src/main/java/com/devmatic/fruits/presenter/viewfruit/ViewFruitContract.kt
@@ -10,6 +10,6 @@ object ViewFruitContract {
     }
 
     interface Presenter : BaseMvpPresenter<View> {
-        fun loadFruit(id: Int?)
+        fun loadFruit(id: Long?)
     }
 }
